fix(app): eslint no-cycle + merging modules with their api to avoid cross dependencies
